requirements.txt
django
pillow
django-ckeditor
gunicorn
django-heroku
python-dotenv
whitenoise
django-summernote
Pillow
